refactor(sdk): align host-contracts-cleartext v0.14 ts layer with v0.13
v0.13 is the source of truth for package structure. v0.14 still carried the pre-split layout: one ~820-line index.ts, with private.ts/public.ts at the top level. It now mirrors v0.13 (and v0.12) exactly — ts/addresses.ts, ts/deploy.ts, ts/types/{private,public}.ts, and an index.ts that is a barrel exposing only deploy, precomputeAddresses and setupACLOwner. Type names follow v0.13 too (InputVerifierInitConfig / KMSVerifierInitConfig / ProtocolConfigInitConfig). No upgrade.ts: no vN->v14 migration exists, same as v0.12. Behaviour is unchanged.
